Rework structure to use less space
Replace interface with embedded types for leafNode and internalNode
with tagging and unsafe type punning. While this is less safe and less
clear the performance wins are worth the additional complexity.

This results in a 14% decrease in leafNode size, 11% decrease in
internalNode size, and a 13% decrease in time per operation.